ai-travel-agent/backup_20250519_134315/requirements.txt
gradio==4.19.2
python-dotenv==1.0.1
requests==2.31.0
reportlab==4.1.0 